+++ packages/observe-sequence/observe_sequence_tests.js
@@ -325,8 +325,19 @@ Tinytest.add('observe sequence - cursor to same cursor', function (test) {
     dep.depend();
     return seq;
   }, function () {
+    coll.insert({_id: "24", rank: 2});
     dep.changed();
-  }, [ {addedAt: ["13", {_id: "13", rank: 1}, 0, null]} ]);
+    coll.insert({_id: "78", rank: 3});
+  }, [
+    {addedAt: ["13", {_id: "13", rank: 1}, 0, null]},
+    {addedAt: ["24", {_id: "24", rank: 2}, 1, null]},
+    {addedAt: ["78", {_id: "78", rank: 3}, 2, null]},
+    // even if the cursor changes to the same cursor, we diff to see if we
+    // missed anything during the invalidation, which leads to these
+    // "changed" events.
+    {changed: ["13", {_id: "13", rank: 1}, {_id: "13", rank: 1}]},
+    {changed: ["24", {_id: "24", rank: 2}, {_id: "24", rank: 2}]}
+  ]);
 });
 
 Tinytest.add('observe sequence - string arrays', function (test) {
@@ -348,5 +359,3 @@ Tinytest.add('observe sequence - string arrays', function (test) {
     {addedAt: ['C', 1, null]}
   ]);
 });
-
-